Refactor udp-notif-pkt into layered architecture (raw/decoded/notification)
Restructure crate with three distinct layers:
- raw: Wire format with unparsed payload (new)
- decoded: Parsed JSON/CBOR notifications
- notification: YANG Push models (moved from yang-push crate)

Breaking changes:
- Move notification types from yang-push to udp-notif-pkt
- Rename packet types: UdpNotifPacket → raw::UdpNotifPacket
- Update imports across collector, pcap-decoder, services, and fuzz

Also adds comprehensive documentation with examples to lib.rs and README.
