+++ crates/udp-notif-service/src/actor.rs
@@ -978,7 +978,7 @@ mod tests {
     use super::*;
     use bytes::{Buf, Bytes, BytesMut};
     use netgauze_parse_utils::WritablePdu;
-    use netgauze_udp_notif_pkt::{MediaType, UdpNotifPacket};
+    use netgauze_udp_notif_pkt::raw::{MediaType, UdpNotifPacket};
     use std::io::Cursor;
     use tokio::net::UdpSocket;
     use tokio::time::Duration;
