+++ crates/yang-push/src/model/mod.rs
@@ -13,6 +13,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod notification;
 pub mod telemetry;
-pub mod udp_notif;
