+++ fuzz/fuzz_targets/fuzz_udp_notif_pkt.rs
@@ -16,7 +16,7 @@
 #![no_main]
 use libfuzzer_sys::fuzz_target;
 use netgauze_parse_utils::{ReadablePdu, Span};
-use netgauze_udp_notif_pkt::UdpNotifPacket;
+use netgauze_udp_notif_pkt::raw::UdpNotifPacket;
 
 fuzz_target!(|data: &[u8]| {
     let mut buf = Span::new(data);
