+++ fuzz/fuzz_targets/fuzz_udp_notif_pkt_serialize.rs
@@ -17,7 +17,7 @@
 
 use libfuzzer_sys::fuzz_target;
 use netgauze_parse_utils::WritablePdu;
-use netgauze_udp_notif_pkt::UdpNotifPacket;
+use netgauze_udp_notif_pkt::raw::UdpNotifPacket;
 use std::io::Cursor;
 
 fuzz_target!(|pkt: UdpNotifPacket| {
